requirements.txt
Django==4.2.5
djangorestframework == 3.15.2
drf-spectacular==0.27.2
django-cors-headers==4.4.0
django-silk==5.2.0
django-filter>=2.4.0
django-environ==0.12.0
psycopg2-binary==2.9.9
